Create javastring_data_randomizer.js
This is a simple javascript code creates random person objects which is in json format. There is a randomizer object which has 4 fields embeded on object as function. When a new person object is created, it gets randomized data from randomizer objects corresponding functions. The requester function is making continues calls to random_indexer fuction. İf a singular request is needed requester can be discarded.